Revert "merge window border and root"
This reverts commit 4ac9e44a7aabf8cb083678a3308842e2fa71a5f5.
